feat: Modify user review page api and rating type 회원 리뷰에서 리뷰를 10개 받던 것에서 4개씩 받을 수 있도록 한다. 별점의 자료형을 int에서 소수점을 받을 수 있게 Double로 수정한다.
Task: #28

+++ server/playbebe/src/main/java/com/knu/server/playbebe/recommend/model/Place.java
@@ -50,7 +50,7 @@ public class Place {
     private List<Review> messages;
 
     //총 별점 계산하기
-    public void setTotalRating(int rating){
+    public void setTotalRating(Double rating){
         if(this.totalRating == null) {
             this.totalRating = Double.valueOf(0);
         }
